Move components higher in the structure
