+++ kolibri/plugins/learn/assets/src/views/LearningActivityDuration/__tests__/index.spec.js
@@ -1,7 +1,7 @@
 import { shallowMount, mount } from '@vue/test-utils';
 
 import { LearningActivities } from 'kolibri.coreVue.vuex.constants';
-import LearningActivityDuration from '../LearningActivityDuration';
+import LearningActivityDuration from '../index';
 
 function makeWrapper(propsData) {
   return mount(LearningActivityDuration, { propsData });
