+++ kolibri/plugins/learn/assets/src/views/LearningActivityDuration/index.vue
@@ -23,7 +23,7 @@
 <script>
 
   import TimeDuration from 'kolibri.coreVue.components.TimeDuration';
-  import useLearningActivities from '../../../composables/useLearningActivities';
+  import useLearningActivities from '../../composables/useLearningActivities';
 
   /**
    * Depending on learning activities and duration of the content node,
